+++ openslp/libslpattr/libslpattr_internal.h
@@ -62,21 +62,20 @@
 
 typedef struct xx_value_t
 {
-    struct xx_value_t *next;
-    int escaped_len;
-    int unescaped_len;
-    union
-    {
-        SLPBoolean va_bool;
-        int va_int;
-        char *va_str; /* This is used for keyword, string, and opaque. */
-    } data; /* Stores the value of the variable. Note, any string must be copied into the struct. */
+   struct xx_value_t * next;
+   size_t escaped_len;
+   size_t unescaped_len;
+   union
+   {
+      SLPBoolean va_bool;
+      int va_int;
+      char * va_str; /* This is used for keyword, string, and opaque. */
+   } data; /* Stores the value of the variable. Note, any string must be copied into the struct. */
 
-    /* Memory handling */
-    struct xx_value_t *next_chunk; /* The next chunk of allocated memory in the value list. */
-    struct xx_value_t *last_value_in_chunk; /* The last value in the chunk. Only set by the chunk head. */
+   /* Memory handling */
+   struct xx_value_t * next_chunk; /* The next chunk of allocated memory in the value list. */
+   struct xx_value_t * last_value_in_chunk; /* The last value in the chunk. Only set by the chunk head. */
 } value_t;
-
 
 /******************************************************************************
  *
@@ -92,15 +91,14 @@ typedef struct xx_value_t
 /* An individual attribute in the struct. */
 typedef struct xx_var_t
 {
-    struct xx_var_t *next; /* Pointer to the next variable. */
-    SLPType type; /* The type of this variable. */
-    const char *tag; /* The name of this variable. */
-    unsigned int tag_len; /* The length of the tag. */
-    value_t *list; /* The list of values. */
-    int list_size; /* The number of values in the list. */
-    SLPBoolean modified; /* Flag. Set to be true if the attribute should be included in the next freshen.  */
+   struct xx_var_t * next; /* Pointer to the next variable. */
+   SLPType type; /* The type of this variable. */
+   const char * tag; /* The name of this variable. */
+   size_t tag_len; /* The length of the tag. */
+   value_t * list; /* The list of values. */
+   int list_size; /* The number of values in the list. */
+   SLPBoolean modified; /* Flag. Set to be true if the attribute should be included in the next freshen.  */
 } var_t;
-
 
 /******************************************************************************
  *
@@ -112,19 +110,19 @@ typedef struct xx_var_t
  */
 struct xx_SLPAttributes
 {
-    SLPBoolean strict; /* Are we using strict typing? */
-    char *lang; /* Language. */
-    var_t *attrs; /* List of vars to be sent. */
-    int attr_count; /* The number of attributes */
+   SLPBoolean strict; /* Are we using strict typing? */
+   char * lang; /* Language. */
+   var_t * attrs; /* List of vars to be sent. */
+   int attr_count; /* The number of attributes */
 };
 
-
-
 /* Finds a variable by its tag. */
-var_t *attr_val_find_str(struct xx_SLPAttributes *slp_attr, const char *tag, int tag_len); 
+var_t * attr_val_find_str(struct xx_SLPAttributes * slp_attr,
+      const char * tag, size_t tag_len); 
 
 /* Finds the type of an attribute. */
-SLPError SLPAttrGetType_len(SLPAttributes attr_h, const char *tag, int tag_len, SLPType *type); 
+SLPError SLPAttrGetType_len(SLPAttributes attr_h, const char * tag,
+      size_t tag_len, SLPType * type); 
 
 /*! @} */
